Add top request table to resource detail page
This was enabled by a substantial refactor of Top.jsx to move the
tapping/websocket code into a self contained module
