+++ web/app/js/components/Top.jsx
@@ -1,14 +1,13 @@
 import _ from 'lodash';
+import { defaultMaxRps } from './util/TapUtils.jsx';
 import ErrorBanner from './ErrorBanner.jsx';
 import PageHeader from './PageHeader.jsx';
-import Percentage from './util/Percentage.js';
 import PropTypes from 'prop-types';
 import React from 'react';
 import TapQueryCliCmd from './TapQueryCliCmd.jsx';
 import TapQueryForm from './TapQueryForm.jsx';
-import TopEventTable from './TopEventTable.jsx';
+import TopModule from './TopModule.jsx';
 import { withContext } from './util/AppContext.jsx';
-import { defaultMaxRps, processTapEvent } from './util/TapUtils.jsx';
 import './../../css/tap.css';
 
 class Top extends React.Component {
@@ -25,8 +24,6 @@ class Top extends React.Component {
     this.loadFromServer = this.loadFromServer.bind(this);
 
     this.state = {
-      tapResultsById: {},
-      topEventIndex: {},
       error: null,
       resourcesByNs: {},
       authoritiesByNs: {},
@@ -41,10 +38,8 @@ class Top extends React.Component {
         authority: "",
         maxRps: defaultMaxRps
       },
-      maxRowsToStore: 40,
-      awaitingWebSocketConnection: false,
-      tapRequestInProgress: false,
       pollingInterval: 10000,
+      tapRequestInProgress: false,
       pendingRequests: false
     };
   }
@@ -54,49 +49,7 @@ class Top extends React.Component {
   }
 
   componentWillUnmount() {
-    if (this.ws) {
-      this.ws.close(1000);
-    }
-    this.stopTapStreaming();
     this.stopServerPolling();
-  }
-
-  onWebsocketOpen = () => {
-    let query = _.cloneDeep(this.state.query);
-    query.maxRps = parseFloat(query.maxRps);
-
-    this.ws.send(JSON.stringify({
-      id: "top-web",
-      ...query
-    }));
-    this.setState({
-      awaitingWebSocketConnection: false,
-      error: null
-    });
-  }
-
-  onWebsocketRecv = e => {
-    this.indexTapResult(e.data);
-  }
-
-  onWebsocketClose = e => {
-    this.stopTapStreaming();
-
-    if (!e.wasClean) {
-      this.setState({
-        error: {
-          error: `Websocket [${e.code}] ${e.reason}`
-        }
-      });
-    }
-  }
-
-  onWebsocketError = e => {
-    this.setState({
-      error: { error: e.message }
-    });
-
-    this.stopTapStreaming();
   }
 
   getResourcesByNs(rsp) {
@@ -127,142 +80,6 @@ class Top extends React.Component {
     };
   }
 
-  parseTapResult = data => {
-    let d = processTapEvent(data);
-
-    if (d.eventType === "responseEnd") {
-      d.latency = parseFloat(d.http.responseEnd.sinceRequestInit.replace("s", ""));
-      d.completed = true;
-    }
-
-    return d;
-  }
-
-  topEventKey = event => {
-    return [event.source.str, event.destination.str, event.http.requestInit.path].join("_");
-  }
-
-  initialTopResult(d, eventKey) {
-    return {
-      count: 1,
-      best: d.responseEnd.latency,
-      worst: d.responseEnd.latency,
-      last: d.responseEnd.latency,
-      success: !d.success ? 0 : 1,
-      failure: !d.success ? 1 : 0,
-      successRate: !d.success ? new Percentage(0, 1) : new Percentage(1, 1),
-      source: d.requestInit.source,
-      sourceLabels: d.requestInit.sourceMeta.labels,
-      destination: d.requestInit.destination,
-      destinationLabels: d.requestInit.destinationMeta.labels,
-      path: d.requestInit.http.requestInit.path,
-      key: eventKey,
-      lastUpdated: Date.now()
-    };
-  }
-
-  incrementTopResult(d, result) {
-    result.count++;
-    if (!d.success) {
-      result.failure++;
-    } else {
-      result.success++;
-    }
-    result.successRate = new Percentage(result.success, result.success + result.failure);
-
-    result.last = d.responseEnd.latency;
-    if (d.responseEnd.latency < result.best) {
-      result.best = d.responseEnd.latency;
-    }
-    if (d.responseEnd.latency > result.worst) {
-      result.worst = d.responseEnd.latency;
-    }
-
-    result.lastUpdated = Date.now();
-  }
-
-  indexTopResult = (d, topResults) => {
-    let eventKey = this.topEventKey(d.requestInit);
-    this.addSuccessCount(d);
-
-    if (!topResults[eventKey]) {
-      topResults[eventKey] = this.initialTopResult(d, eventKey);
-    } else {
-      this.incrementTopResult(d, topResults[eventKey]);
-    }
-
-    if (_.size(topResults) > this.state.maxRowsToStore) {
-      this.deleteOldestIndexedResult(topResults);
-    }
-
-    return topResults;
-  }
-
-  indexTapResult = data => {
-    // keep an index of tap results by id until the request is complete.
-    // when the request has completed, add it to the aggregated Top counts and
-    // discard the individual tap result
-    let resultIndex = this.state.tapResultsById;
-    let d = this.parseTapResult(data);
-
-    if (_.isNil(resultIndex[d.id])) {
-      // don't let tapResultsById grow unbounded
-      if (_.size(resultIndex) > this.state.maxRowsToStore) {
-        this.deleteOldestIndexedResult(resultIndex);
-      }
-
-      resultIndex[d.id] = {};
-    }
-    resultIndex[d.id][d.eventType] = d;
-
-    // assumption: requests of a given id all share the same high level metadata
-    resultIndex[d.id].base = d;
-    resultIndex[d.id].lastUpdated = Date.now();
-
-    let topIndex = this.state.topEventIndex;
-    if (d.completed) {
-      // only add results into top if the request has completed
-      // we can also now delete this result from the Tap result index
-      topIndex = this.indexTopResult(resultIndex[d.id], topIndex);
-      delete resultIndex[d.id];
-    }
-
-    this.setState({
-      tapResultsById: resultIndex,
-      topEventIndex: topIndex
-    });
-  }
-
-  addSuccessCount = d => {
-    // cope with the fact that gRPC failures are returned with HTTP status 200
-    // and correctly classify gRPC failures as failures
-    let success = parseInt(d.responseInit.http.responseInit.httpStatus, 10) < 500;
-    if (success) {
-      let grpcStatusCode = _.get(d, "responseEnd.http.responseEnd.eos.grpcStatusCode");
-      if (!_.isNil(grpcStatusCode)) {
-        success = grpcStatusCode === 0;
-      } else if (!_.isNil(_.get(d, "responseEnd.http.responseEnd.eos.resetErrorCode"))) {
-        success = false;
-      }
-    }
-
-    d.success = success;
-  }
-
-  deleteOldestIndexedResult = resultIndex => {
-    let oldest = Date.now();
-    let oldestId = "";
-
-    _.each(resultIndex, (res, id) => {
-      if (res.lastUpdated < oldest) {
-        oldest = res.lastUpdated;
-        oldestId = id;
-      }
-    });
-
-    delete resultIndex[oldestId];
-  }
-
   startServerPolling() {
     this.loadFromServer();
     this.timerId = window.setInterval(this.loadFromServer, this.state.pollingInterval);
@@ -271,40 +88,6 @@ class Top extends React.Component {
   stopServerPolling() {
     window.clearInterval(this.timerId);
     this.api.cancelCurrentRequests();
-  }
-
-  startTapSteaming() {
-    this.setState({
-      awaitingWebSocketConnection: true,
-      tapRequestInProgress: true,
-      tapResultsById: {},
-      topEventIndex: {}
-    });
-
-    let protocol = window.location.protocol === "https:" ? "wss" : "ws";
-    let tapWebSocket = `${protocol}://${window.location.host}${this.props.pathPrefix}/api/tap`;
-
-    this.ws = new WebSocket(tapWebSocket);
-    this.ws.onmessage = this.onWebsocketRecv;
-    this.ws.onclose = this.onWebsocketClose;
-    this.ws.onopen = this.onWebsocketOpen;
-    this.ws.onerror = this.onWebsocketError;
-  }
-
-  stopTapStreaming() {
-    this.setState({
-      tapRequestInProgress: false,
-      awaitingWebSocketConnection: false
-    });
-  }
-
-  handleTapStart = e => {
-    e.preventDefault();
-    this.startTapSteaming();
-  }
-
-  handleTapStop = () => {
-    this.ws.close(1000);
   }
 
   loadFromServer() {
@@ -347,29 +130,39 @@ class Top extends React.Component {
     });
   }
 
-  render() {
-    let tableRows = _.values(this.state.topEventIndex);
+  handleTapStart = () => {
+    this.setState({
+      tapRequestInProgress: true
+    });
+  }
 
+  handleTapStop = () => {
+    this.setState({
+      tapRequestInProgress: false
+    });
+  }
+
+  render() {
     return (
       <div>
         {!this.state.error ? null :
         <ErrorBanner message={this.state.error} onHideMessage={() => this.setState({ error: null })} />}
-
         <PageHeader header="Top" />
         <TapQueryForm
           enableAdvancedForm={false}
-          tapRequestInProgress={this.state.tapRequestInProgress}
-          awaitingWebSocketConnection={this.state.awaitingWebSocketConnection}
           handleTapStart={this.handleTapStart}
           handleTapStop={this.handleTapStop}
           resourcesByNs={this.state.resourcesByNs}
           authoritiesByNs={this.state.authoritiesByNs}
+          tapRequestInProgress={this.state.tapRequestInProgress}
           updateQuery={this.updateQuery}
           query={this.state.query} />
 
         <TapQueryCliCmd cmdName="top" query={this.state.query} />
-
-        <TopEventTable tableRows={tableRows} />
+        <TopModule
+          pathPrefix={this.props.pathPrefix}
+          query={this.state.query}
+          startTap={this.state.tapRequestInProgress} />
       </div>
     );
   }
